@@ -1,3 +1,3 @@
 module github.com/kvark128/dodp
 
-go 1.16
+go 1.20
